+++ src/ICESat-2/ICESat-2.jl
@@ -24,7 +24,7 @@ Base.@kwdef mutable struct ICESat2_Granule{product} <: Granule
 end
 
 function Base.copy(g::ICESat2_Granule{product}) where {product}
-    ICESat2_Granule{product}(g.id, g.url, copy(g.bbox), copy(g.info), copy(g.polygons))
+    ICESat2_Granule{product}(g.id, g.url, g.info, copy(g.polygons))
 end
 
 """
@@ -53,7 +53,6 @@ function bounds(granule::ICESat2_Granule)
             max_x = nt.eastBoundLongitude,
             max_y = nt.northBoundLatitude,
         )
-        ntb
     end
 end